Four-way-movement component. Added `elapsed` timestamp all the way from game to components so you know how far to move Converted `Entity` coordinates to floats, ints don't work with high-speed updates or low velocities otherwise Added missing `EntityExtensions` documentation Replaced tester sprite with `Charspore`

+++ Puffin.Core/Ecs/Entity.cs
@@ -12,10 +12,10 @@ namespace Puffin.Core.Ecs
     public class Entity
     {
         private Dictionary<Type, Component> components = new Dictionary<Type, Component>();
-        private int x = 0;
-        private int y = 0;
+        private float x = 0;
+        private float y = 0;
 
-        public int X {
+        public float X {
             get { return x; }
             set { 
                 x = value;
@@ -23,7 +23,7 @@ namespace Puffin.Core.Ecs
             }
         }
 
-        public int Y {
+        public float Y {
             get { return y; }
             set { 
                 y = value;
